Add SemVer compatibility checks (#1556)
This integrates the cargo-semver-checks tool in the CI to check for
SemVer compatibility as per the Cargo reference
(https://doc.rust-lang.org/cargo/reference/semver.html).

The cargo-public-api tool is a viable alternative to
cargo-semver-checks, but I didn't choose it because its comparison
baseline needs to be stored in the repository and manually updated on
each change. For zenoh, the 1.0.0 revision will be the baseline for the
forseeable future.

This comes with a caveat though: cargo-semver-checks accepts a "release
type" argument which would be either patch or minor in zenoh's case. I
specified "patch" since I expect patch releases to be more common. If
the next release is expected to include changes that require a minor
version bump (as defined by the Cargo reference) then the CI workflow
would need to be manually adjusted at the start and the end of the release cycle.