avoid copying shm buffer when sending it on network (#2062)
* avoid copying shm buffer when sending it on network

The buffer copy was useless, as the content of the buffer was written
anyway.

The increment of the shm refcount has also been moved to the
serialization step, as it avoids incrementing uselessly in case of
message being dropped (even if the watchdog catch it eventually).

Last but not least, the code has been refactored to remove macros.

* style

* oupsie

* fix info serialization