Rework `HatInterestTrait`
The old API didn’t allow for in-region interest propagation because it
assumed that interest only originates from south faces, which is untrue
in router regions.

We are still missing Declare extensions to properly route current
declarations to the interest source in router subregions. The current
plan is to stop design of the interests interface at this stage and drop
support for the interest protocol in router regions so we can prioritize
other concerns. The downside is that router regions can only be deployed
at the backbone, at a subregion of the backbone, or at the leaf regions.