stage0: Base
- Propagate `self: &HatCode` in hat call tree

`HatCode` strucs are currently empty, which makes it hard to introduce
multiple hats within a runtime router, as there is no way to distinguish
hats.

- Move `HatTables` into `HatCode`

This patch splits the Tables struct into data (shared data) and hat and
moves hat logic and trait objects out of TablesData. Note that it's
always possible to mutably borrow both the hat and the tables because
they're two different struct fields.

- Remove `linkstate_peer` network and hat

This change also removes peer failover brokering in the router hat and
scraps the router election mechanism used to avoid duplicate PUSHs. The
rationale behind this is that the reduced router hat would rely on a
linkstate-enabled peer hat to fulfill these functionalities.

- Add handshake remote bound extension

This patch informs a node of its bound in a remote gateway.

This commit also removes peer failover brokering in the router hat and
scraps the router election mechanism used to avoid duplicate PUSHs. The
rationale behind this is that the reduced router hat would rely on a
linkstate-enabled peer hat to fulfill these functionalities.

Co-Authored-By: OlivierHecart <olivier.hecart@adlinktech.com>